+++ adapters/constants/index.js
@@ -12,15 +12,15 @@ exports.AREA = 'area';
 exports.POINT = 'point';
 exports.LINE = 'line';
 
-exports.ID_DEFAULTS = { point: ['point'], line: ['line'], area: ['area'], vertex: ['vertex'] };
+exports.ID_DEFAULTS = { point: ['point'], line: ['line'], area: ['area'], vertex: ['vertex'], relation: ['relation'] };
 
 // default geometry iD presets!
 exports.ID_GEOM_PRESETS = {
-    point: { fields: ['name'], geometry: ['point'], tags: {}, name: 'Point', matchScore: 0.1},
+    point: { fields: ['name'], geometry: ['point'], tags: {}, name: 'Point', matchScore: 0.1} ,
     line: { fields: ['name'], geometry: ['line'], tags: {}, name: 'Line', matchScore: 0.1},
     area: { fields: ['name'], geometry: ['area'], tags: { 'area': 'yes' }, name: 'Area', matchScore: 0.1},
     vertex: { fields: [ 'name' ], geometry: [ 'vertex' ], tags: {}, name: 'Other', matchScore: 0.1},
-    relation: { icon: 'iD-relation', fields: [ 'name', 'relation' ], geometry: [ 'relation' ], tags: {}, name: 'Relation' }
+    relation: { icon: 'iD-relation', fields: [ 'name', 'relation' ], geometry: [ 'relation' ], tags: {}, name: 'Relation', matchScore: 0.1}
 };
 exports.ID_GENERIC_FIELDS = {
     name: {
@@ -62,5 +62,26 @@ exports.ID_GENERIC_FIELDS = {
         type: 'semiCombo',
         label: 'Suggested Hashtags',
         placeholder: '#example'
+    }
+};
+
+exports.ID_DEFAULT_CATEGORIES = {
+    'category-area': {
+        icon: 'maki-natural',
+        geometry: 'area',
+        name: 'MapRules Area Features',
+        members: ['area']
+    },
+    'category-point': {
+        icon: 'maki-natural',
+        geometry: 'point',
+        name: 'MapRules Point Features',
+        members: ['point']
+    },
+    'category-line': {
+        icon: 'maki-natural',
+        geometry: 'line',
+        name: 'MapRules Point Features',
+        members: ['line']
     }
 };