+++ test/presetConfig/handlersSpec.js
@@ -29,14 +29,14 @@ module.exports = () => {
         it('replies 200 and presetConfig provided uuid that exists in the db', async () => {
             try {
                 const request = mergeDefaults({ 
-                        method: 'GET',
-                        url: `/config/${seedId}`
-                    }),
-                    r = await server.inject(request),
-                    statusCode = r.statusCode,
-                    presetConfig = r.result,
-                    validation = Joi.validate(presetConfig,presetConfigSchema);
-                expect(r.statusCode).to.equal(200);
+                    method: 'GET',
+                    url: `/config/${seedId}`
+                });
+                const r = await server.inject(request);
+                const statusCode = r.statusCode;
+                const presetConfig = r.result;
+                const validation = Joi.validate(presetConfig, presetConfigSchema);
+                expect(statusCode).to.equal(200);
                 expect(validation.error).to.be.null;
               
             } catch (error) {
@@ -47,10 +47,10 @@ module.exports = () => {
         it('replies 400 if uuid is invalid', async () => {
             try {
                 const request = mergeDefaults({ 
-                        method: 'GET',
-                        url: `/config/${uuidv1()}`
-                    }), 
-                    r = await server.inject(request);
+                    method: 'GET',
+                    url: `/config/${uuidv1()}`
+                });
+                const r = await server.inject(request);
 
                 expect(r.statusCode).to.equal(400);
 
@@ -61,11 +61,11 @@ module.exports = () => {
         it('replies 404 if the the uuid is not in the db', async () => {
             try {
                 const request = mergeDefaults({
-                        method: 'GET',
-                        url: `/config/${uuidv4()}`
-                    }),
-                    r = await server.inject(request),
-                    statusCode = r.statusCode;
+                    method: 'GET',
+                    url: `/config/${uuidv4()}`
+                });
+                const r = await server.inject(request);
+                const statusCode = r.statusCode;
 
                 expect(r.statusCode).to.equal(404);
             } catch (error) {
@@ -80,11 +80,11 @@ module.exports = () => {
             validPresetConfigs.forEach(async (validPresetConfig) => {
                 try {
                     const request = mergeDefaults({
-                            method: 'PUT',
-                            payload: validPresetConfig,
-                            url: `/config/${seedId}`
-                        }),
-                        r = await server.inject(request);
+                        method: 'PUT',
+                        payload: validPresetConfig,
+                        url: `/config/${seedId}`
+                    });
+                    const r = await server.inject(request);
                         
                     expect(r.statusCode).to.equal(200);
                 
@@ -98,13 +98,13 @@ module.exports = () => {
             invalidPresetConfigs.forEach(async (invalidPresetConfig) => {
                 try {
                     const request = mergeDefaults({
-                            method: 'PUT',
-                            payload: invalidPresetConfig,
-                            url: `/config/${seedId}`
-                        }),
-                        r = await server.inject(request),
-                        statusCode = r.statusCode,
-                        message = r.result.message;
+                        method: 'PUT',
+                        payload: invalidPresetConfig,
+                        url: `/config/${seedId}`
+                    });
+                    const r = await server.inject(request);
+                    const statusCode = r.statusCode;
+                    const message = r.result.message;
 
                     expect(statusCode).to.be.equal(400);
                     expect(message).to.not.be.null;
@@ -119,13 +119,13 @@ module.exports = () => {
             invalidPresetConfigs.forEach(async (invalidPresetConfig) => {
                 try {
                     const request = mergeDefaults({
-                            method: 'PUT',
-                            payload: invalidPresetConfig,
-                            url: `/config/${uuidv1()}`
-                        }),
-                        r = await server.inject(request),
-                        statusCode = r.statusCode,
-                        message = r.result.message;
+                        method: 'PUT',
+                        payload: invalidPresetConfig,
+                        url: `/config/${uuidv1()}`
+                    });
+                    const r = await server.inject(request);
+                    const statusCode = r.statusCode;
+                    const message = r.result.message;
 
                     expect(statusCode).to.be.equal(400);
                     expect(message).to.not.be.null;
@@ -140,12 +140,12 @@ module.exports = () => {
             validPresetConfigs.forEach(async (validPresetConfig) => {
                 try {
                     const request = mergeDefaults({
-                            method: 'PUT',
-                            payload: validPresetConfig,
-                            url: `/config/${uuidv4()}`
-                        }),
-                        r = await server.inject(request),
-                        statusCode = r.statusCode;
+                        method: 'PUT',
+                        payload: validPresetConfig,
+                        url: `/config/${uuidv4()}`
+                    });
+                    const r = await server.inject(request);
+                    const statusCode = r.statusCode;
 
                     expect(statusCode).to.be.equal(404);
 
@@ -160,12 +160,12 @@ module.exports = () => {
         it ('replies 200 if provided presetConig is valid', async () => {
             validPresetConfigs.forEach(async (validPresetConfig) => {
                 const request = mergeDefaults({
-                        method: 'post',
-                        payload: validPresetConfig,
-                        url: '/config'
-                    }),
-                    r = await server.inject(request),
-                    statusCode = r.statusCode;
+                    method: 'post',
+                    payload: validPresetConfig,
+                    url: '/config'
+                });
+                const r = await server.inject(request);
+                const statusCode = r.statusCode;
 
                 expect(statusCode).to.equal(200);
             });
@@ -173,12 +173,12 @@ module.exports = () => {
         it('replies 400 if provided presetConfig is invalid', async () => {
             invalidPresetConfigs.forEach(async (invalidPresetConfig) => {
                 const request = mergeDefaults({
-                        method: 'post',
-                        payload: invalidPresetConfig,
-                        url: '/config'
-                    }),
-                    r = await server.inject(request),
-                    statusCode = r.statusCode;
+                    method: 'post',
+                    payload: invalidPresetConfig,
+                    url: '/config'
+                });
+                const r = await server.inject(request);
+                const statusCode = r.statusCode;
 
                 expect(statusCode).to.equal(400);
             });
@@ -187,18 +187,17 @@ module.exports = () => {
             invalidPresetConfigs.forEach(async (invalidPresetConfig) => {
                 try {
                     const request = mergeDefaults({
-						  	  method: 'post',
-							  payload: invalidPresetConfig,
-							  url: '/config'
-						  }),
-						  r = await server.inject(request),
-                        message = r.result.message;
+                        method: 'post',
+                        payload: invalidPresetConfig,
+                        url: '/config'
+                    });
+                    const r = await server.inject(request);
+                    const message = r.result.message;
                     
                     expect(message).to.not.be.null;
 
     			} catch (error) {
                     console.error(error);
-
                 }
             });
         });