+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 0.0.3
+
+##### December 14th, 2018
+
+Bugs/Fixes :bug:
+
+- documentation updates [#14] [#19]
+
+Enhancements :star2:
+
+- settle on 'point', 'line', and 'area' as the default geometry types [#16]
+- increase iD presets' [matchScore](https://github.com/openstreetmap/iD/blob/master/data/presets/schema/preset.json#L74) [#24]
+
 # 0.0.2-alpha
 
 ##### October 12th, 2018
@@ -18,4 +31,7 @@ Bugs/Fixes :bug:
 [#1]: https://github.com/radiant-maxar/maprules/issues/1
 [#4]: https://github.com/radiant-maxar/maprules/issues/4
 [#5]: https://github.com/radiant-maxar/maprules/issues/5
-
+[#14]: https://github.com/radiant-maxar/maprules/issues/14
+[#16]: https://github.com/radiant-maxar/maprules/issues/16
+[#19]: https://github.com/radiant-maxar/maprules/issues/19
+[#24]: https://github.com/radiant-maxar/maprules/issues/24